split AllBlocks to BlockBuilder ModelBuilder StateBuilder
