+++ src/main/java/shagejack/industrimania/content/worldGen/OreParagenesisRegistry.java
@@ -1,7 +1,6 @@
 package shagejack.industrimania.content.worldGen;
 
 import shagejack.industrimania.content.worldGen.record.Ore;
-import shagejack.industrimania.registers.AllBlocks;
 
 public class OreParagenesisRegistry {
     //the paragenesis of ores doesn't need availableRock, minY and maxY, but chanceAsParagenesis is necessary
