+++ src/main/java/shagejack/industrimania/content/worldGen/OreRegistry.java
@@ -3,9 +3,8 @@ package shagejack.industrimania.content.worldGen;
 import com.google.common.collect.Lists;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.Blocks;
-import net.minecraftforge.registries.RegistryObject;
 import shagejack.industrimania.content.worldGen.record.Ore;
-import shagejack.industrimania.registers.AllBlocks;
+import shagejack.industrimania.registers.block.AllBlocks;
 
 import java.util.List;
